Build WebKit as C++11 on Mac https://bugs.webkit.org/show_bug.cgi?id=100720
Reviewed by Daniel Bates.

Source/JavaScriptCore:

* Configurations/Base.xcconfig:
Add CLANG_CXX_LANGUAGE_STANDARD=gnu++0x.

* bytecompiler/BytecodeGenerator.cpp:
(JSC::BytecodeGenerator::generate):
(JSC::BytecodeGenerator::pushFinallyContext):
(JSC::BytecodeGenerator::beginSwitch):
* llint/LLIntOffsetsExtractor.cpp:
* runtime/Identifier.cpp:
(JSC::Identifier::add8):
* runtime/Identifier.h:
(JSC::Identifier::add):
* runtime/JSONObject.cpp:
(JSC::appendStringToStringBuilder):
* runtime/StringPrototype.cpp:
(JSC::replaceUsingStringSearch):
Add static_casts to prevent implicit type conversions in non-constant initializer lists.

Source/ThirdParty/ANGLE:

* Configurations/Base.xcconfig:
Add CLANG_CXX_LANGUAGE_STANDARD=gnu++0x.

Source/WebCore:

* Configurations/Base.xcconfig:
Add CLANG_CXX_LANGUAGE_STANDARD=gnu++0x.

* platform/graphics/FontPlatformData.h:
(WebCore::FontPlatformData::hash):
* platform/graphics/cg/GraphicsContextCG.cpp:
(WebCore::GraphicsContext::drawLine):
(WebCore::GraphicsContext::strokeArc):
* platform/graphics/filters/CustomFilterProgramInfo.cpp:
(WebCore::CustomFilterProgramInfo::hash):
(WebCore::ProtectionSpaceHash::hash):
* platform/network/cf/FormDataStreamCFNet.cpp:
(WebCore::setHTTPBody):
Add static_casts to prevent implicit type conversions in non-constant initializer lists.

* platform/mac/ClipboardMac.mm:
(WebCore::ClipboardMac::createDragImage):
* platform/mac/CursorMac.mm:
(WebCore::createNamedCursor):
* platform/mac/ScrollViewMac.mm:
(WebCore::ScrollView::platformSetContentsSize):
Use NSMakePoint to prevent implicit type conversions in non-constant initializer lists.

* platform/mac/CookieJar.mm:
(WebCore::cookiesEnabled):
* platform/network/mac/ResourceHandleMac.mm:
(WebCore::shouldRelaxThirdPartyCookiePolicy):
Add explicit casts to NSHTTPCookieAcceptPolicy.

* platform/mac/WebCoreNSURLExtras.h:
* platform/mac/WebCoreNSURLExtras.mm:
(WebCore::URLByTruncatingOneCharacterBeforeComponent):
(WebCore::dataForURLComponentType):
* platform/network/ProtectionSpaceHash.h:
Use CFIndex instead of CFURLComponentType, the enum can be mangled differently in some cases, and these
functions are exported from WebCore.

Source/WebKit/mac:

* Configurations/Base.xcconfig:
Add CLANG_CXX_LANGUAGE_STANDARD=gnu++0x.

* History/WebBackForwardList.mm:
(-[WebBackForwardList description]):
* History/WebHistoryItem.mm:
(-[WebHistoryItem description]):
Add static_casts to prevent implicit type conversions in non-constant initializer lists.

Source/WebKit2:

* Configurations/Base.xcconfig:
Add CLANG_CXX_LANGUAGE_STANDARD=gnu++0x.

* UIProcess/mac/WebCookieManagerProxyMac.mm:
(WebKit::WebCookieManagerProxy::persistHTTPCookieAcceptPolicy):
* WebProcess/Cookies/mac/WebCookieManagerMac.mm:
(WebKit::WebCookieManager::platformSetHTTPCookieAcceptPolicy):
Add explicit casts to NSHTTPCookieAcceptPolicy.

Source/WTF:

* Configurations/Base.xcconfig:
Add CLANG_CXX_LANGUAGE_STANDARD=gnu++0x.

* wtf/text/AtomicString.cpp:
(WTF::AtomicString::add):
Change the type of length to be unsigned to prevent implicit type conversions in non-constant initializer lists.

* wtf/unicode/UTF8.cpp:
(Unicode):
Add static_casts to prevent implicit type conversions in non-constant initializer lists.

Tools:

* DumpRenderTree/mac/Configurations/Base.xcconfig:
* MiniBrowser/Configurations/Base.xcconfig:
* TestWebKitAPI/Configurations/Base.xcconfig:
* WebKitTestRunner/Configurations/Base.xcconfig:
Add CLANG_CXX_LANGUAGE_STANDARD=gnu++0x.

Canonical link: https://commits.webkit.org/118809@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132925 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/JavaScriptCore/bytecompiler/BytecodeGenerator.cpp
@@ -170,8 +170,8 @@ JSObject* BytecodeGenerator::generate()
         TryRange& range = m_tryRanges[i];
         ASSERT(range.tryData->targetScopeDepth != UINT_MAX);
         HandlerInfo info = {
-            range.start->bind(0, 0), range.end->bind(0, 0),
-            range.tryData->target->bind(0, 0), range.tryData->targetScopeDepth
+            static_cast<uint32_t>(range.start->bind(0, 0)), static_cast<uint32_t>(range.end->bind(0, 0)),
+            static_cast<uint32_t>(range.tryData->target->bind(0, 0)), static_cast<uint32_t>(range.tryData->targetScopeDepth)
 #if ENABLE(JIT)
             ,
 #if ENABLE(LLINT)
@@ -2131,11 +2131,11 @@ void BytecodeGenerator::pushFinallyContext(StatementNode* finallyBlock)
     scope.isFinallyBlock = true;
     FinallyContext context = {
         finallyBlock,
-        m_scopeContextStack.size(),
-        m_switchContextStack.size(),
-        m_forInContextStack.size(),
-        m_tryContextStack.size(),
-        m_labelScopes.size(),
+        static_cast<unsigned>(m_scopeContextStack.size()),
+        static_cast<unsigned>(m_switchContextStack.size()),
+        static_cast<unsigned>(m_forInContextStack.size()),
+        static_cast<unsigned>(m_tryContextStack.size()),
+        static_cast<unsigned>(m_labelScopes.size()),
         m_finallyDepth,
         m_dynamicScopeDepth
     };
@@ -2462,7 +2462,7 @@ void BytecodeGenerator::emitPushNameScope(const Identifier& property, RegisterID
 
 void BytecodeGenerator::beginSwitch(RegisterID* scrutineeRegister, SwitchInfo::SwitchType type)
 {
-    SwitchInfo info = { instructions().size(), type };
+    SwitchInfo info = { static_cast<uint32_t>(instructions().size()), type };
     switch (type) {
         case SwitchInfo::SwitchImmediate:
             emitOpcode(op_switch_imm);
