+++ Source/WebKit/mac/History/WebHistoryItem.mm
@@ -232,7 +232,7 @@ void WKNotifyHistoryItemChanged(HistoryItem*)
             [result appendString:[child description]];
         }
         // shift all the contents over.  A bit slow, but hey, this is for debugging.
-        NSRange replRange = {currPos, [result length]-currPos};
+        NSRange replRange = { static_cast<NSUInteger>(currPos), [result length] - currPos };
         [result replaceOccurrencesOfString:@"\n" withString:@"\n    " options:0 range:replRange];
     }
